# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PocketChange therapy practice management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients/{code}/portal:
    get:
      operationId: getClientPortal
      tags: [portal]
      summary: Get client portal data
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Portal data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalData"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients/{code}/goals:
    get:
      operationId: getClientGoals
      tags: [goals]
      summary: Get all goals for a client
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Goals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal for a client
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /clients/{code}/goals/{goalId}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalBody"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /client-portal/sms-reminder:
    get:
      operationId: getSmsReminder
      tags: [reminders]
      summary: Get SMS reminder settings
      responses:
        "200":
          description: SMS reminder settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsReminder"
        "404":
          description: Not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveSmsReminder
      tags: [reminders]
      summary: Save SMS reminder settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsReminderBody"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsReminder"
    delete:
      operationId: deleteSmsReminder
      tags: [reminders]
      summary: Delete SMS reminder settings
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        accessCode:
          type: string
        notes:
          type: ["string", "null"]
        practitionerId:
          type: integer
        createdAt:
          type: string
          format: date-time
        goalCount:
          type: ["integer", "null"]
        lastCheckin:
          type: ["string", "null"]
        averageProgress:
          type: ["number", "null"]
      required:
        - id
        - name
        - accessCode
        - practitionerId
        - createdAt
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          enum: [process, outcome, null]
        parentGoalId:
          type: ["integer", "null"]
        clientId:
          type: integer
        currentProgress:
          type: ["number", "null"]
        targetDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        subGoals:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
      required:
        - id
        - title
        - clientId
        - createdAt
        - subGoals
    CreateGoalBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          enum: [process, outcome, null]
        parentGoalId:
          type: ["integer", "null"]
        targetDate:
          type: ["string", "null"]
      required:
        - title
    UpdateGoalBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          enum: [process, outcome, null]
        targetDate:
          type: ["string", "null"]
...
    PortalData:
      type: object
      properties:
        client:
          $ref: "#/components/schemas/Client"
        practitionerName:
          type: string
        goals:
          type: array
          items:
            $ref: "#/components/schemas/Goal"
      required:
        - client
        - practitionerName
        - goals
...
    SmsReminder:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        days:
          type: string
        time:
          type: string
        timezone:
          type: string
        messageTemplate:
          type: string
      required:
        - id
        - phone
        - days
        - time
        - timezone
        - messageTemplate
    SmsReminderBody:
      type: object
      properties:
        phone:
          type: string
        days:
          type: string
        time:
          type: string
        timezone:
          type: string
        messageTemplate:
          type: string
      required:
        - phone
        - days
        - time
        - timezone
        - messageTemplate