# lib/api-zod/src/generated/types/createGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PocketChange therapy practice management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGoalBodyType } from "./createGoalBodyType";

export interface CreateGoalBody {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  type?: CreateGoalBodyType;
  /** @nullable */
  parentGoalId?: number | null;
  /** @nullable */
  targetDate?: string | null;
}