# lib/api-zod/src/generated/types/updateGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PocketChange therapy practice management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateGoalBodyType } from "./updateGoalBodyType";

export interface UpdateGoalBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  type?: UpdateGoalBodyType;
  /** @nullable */
  targetDate?: string | null;
}